Extract code dealing with ProbeSets to their own module.
